添加了文章操作
添加了文章操作

+++ src/main/java/xyz/oilpea/schooldemo/demo/dao/CarouselMapper.java
@@ -3,13 +3,13 @@ package xyz.oilpea.schooldemo.demo.dao;
 import xyz.oilpea.schooldemo.demo.pojo.Carousel;
 
 public interface CarouselMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Carousel record);
 
     int insertSelective(Carousel record);
 
-    Carousel selectByPrimaryKey(Integer id);
+    Carousel selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Carousel record);
