+++ src/main/java/xyz/oilpea/schooldemo/demo/dao/CompetitionMapper.java
@@ -3,13 +3,13 @@ package xyz.oilpea.schooldemo.demo.dao;
 import xyz.oilpea.schooldemo.demo.pojo.Competition;
 
 public interface CompetitionMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Competition record);
 
     int insertSelective(Competition record);
 
-    Competition selectByPrimaryKey(Integer id);
+    Competition selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Competition record);
