+++ src/main/java/xyz/oilpea/schooldemo/demo/dao/LaureateMapper.java
@@ -3,13 +3,13 @@ package xyz.oilpea.schooldemo.demo.dao;
 import xyz.oilpea.schooldemo.demo.pojo.Laureate;
 
 public interface LaureateMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Laureate record);
 
     int insertSelective(Laureate record);
 
-    Laureate selectByPrimaryKey(Integer id);
+    Laureate selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Laureate record);
