+++ src/main/java/xyz/oilpea/schooldemo/demo/dao/WinMapper.java
@@ -3,13 +3,13 @@ package xyz.oilpea.schooldemo.demo.dao;
 import xyz.oilpea.schooldemo.demo.pojo.Win;
 
 public interface WinMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Win record);
 
     int insertSelective(Win record);
 
-    Win selectByPrimaryKey(Integer id);
+    Win selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Win record);
